convert methods in every store

+++ Phase 2/MoTraCoCa/src/main/java/com/motracoca/model/Price.java
@@ -1,4 +1,4 @@
 package com.motracoca.model;
 
-public record Price(int price) {
+public record Price(double price) {
 }
